+++ src/js/shop/components/Checkout/index.js
@@ -1,5 +1,5 @@
 export {default as Checkout} from './Checkout';
 export {default as Account} from './Account';
-export {default as Address} from './Address';
+export {default as Delivery} from './Delivery';
 export {default as Payment} from './Payment';
 export {default as Confirmation} from './Confirmation';
